OSIS-2203 Queryset on learning_unit instead of learning_unit_year
